Add protobuf migration and reverse engineering TODOs
Changes:
- Add REVERSE_ENGINEERING_TODOS.md documenting all incomplete RE work
- Add PROTOBUF_MIGRATION.md documenting protobuf stub replacement
- Create tunnel.proto with proper protobuf definitions
- Add helpers.go for GetHeaders() compatibility
- Add verify_proto_test.go to verify ProtoReflect works
- Update MODULE.bazel with rules_proto dependency
- Update tunnelpb BUILD.bazel for proto generation
- Remove types.go stub (replaced by generated code)

Note: Uses rules_proto (deprecated) - should migrate to modern
bazel_dep(name = "protobuf") in future iteration.

Background agent a072473 working on codesign/MCP server stubs.
