+++ claude_web_env/re/environment_manager/6b49f1ca/src/internal/tunnel/tunnelpb/BUILD.bazel
@@ -1,11 +1,30 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_library")
+load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
+load("@io_bazel_rules_go//proto:def.bzl", "go_proto_library")
+load("@rules_proto//proto:defs.bzl", "proto_library")
+
+proto_library(
+    name = "tunnel_proto",
+    srcs = ["tunnel.proto"],
+    visibility = ["//visibility:public"],
+)
+
+go_proto_library(
+    name = "tunnel_go_proto",
+    importpath = "github.com/anthropics/anthropic/api-go/environment-manager/internal/tunnel/tunnelpb",
+    proto = ":tunnel_proto",
+    visibility = ["//visibility:public"],
+)
 
 go_library(
     name = "tunnelpb",
-    srcs = ["types.go"],
+    srcs = ["helpers.go"],
+    embed = [":tunnel_go_proto"],
     importpath = "github.com/anthropics/anthropic/api-go/environment-manager/internal/tunnel/tunnelpb",
     visibility = ["//claude_web_env/re/environment_manager/6b49f1ca/src:__subpackages__"],
-    deps = [
-        "@org_golang_google_protobuf//reflect/protoreflect",
-    ],
+)
+
+go_test(
+    name = "tunnelpb_test",
+    srcs = ["verify_proto_test.go"],
+    embed = [":tunnelpb"],
 )
